pyproject.toml
[tool.poetry]
name = "children-game-night"
version = "0.1.0"
description = ""
authors = ["Jonathan Zhang <jonathan.zhang@scale.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
pyserial = "^3.5"
flask-socketio = "^5.3.6"
pygame = "^2.5.2"
flask-cors = "^4.0.0"
ruff = "^0.7.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
types-pyserial = "^3.5.0.10"
types-flask-cors = "^4.0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
